+++ CHANGELOG.md
@@ -31,7 +31,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- `deal.show()` added `tasks: [...taskids]` key to the resolved value
+- `deal.show()` added key `tasks: { [idx]: [taskid] }` to the resolved value
 
 ### Changed
 
@@ -41,7 +41,11 @@ All notable changes to this project will be documented in this file.
 - dependencies update
 - [Breaking] `iexec deal show` ends with error when the deal doesn't exists
 - [Breaking] `deal.show()` throw when the deal doesn't exists
-- [Breaking] `deal.computeTaskIdsArray()` is no longer exposedin the lib (`deal.show()` resolves now as `{..., tasks: [...taskids]}`)
+- [Breaking] `deal.computeTaskIdsArray()` is no longer exposed (`deal.show()` resolves now as `{..., tasks: { [idx]: [taskid] }}`)
+- [Breaking] `order.signOrder()` is replaced by dedicated methods `order.signApporder()`, `order.signDatasetorder()`, `order.signWorkerpoolorder()`, `order.signRequestorder()`
+- [Breaking] `order.cancelOrder()` is replaced by dedicated methods `order.cancelApporder()`, `order.cancelDatasetorder()`, `order.cancelWorkerpoolorder()`, `order.cancelRequestorder()`
+- [Breaking] `order.publishOrder()` is replaced by dedicated methods `order.publishApporder()`, `order.publishDatasetorder()`, `order.publishWorkerpoolorder()`, `order.publishRequestorder()`
+- [Breaking] `order.unpublishOrder()` is replaced by dedicated methods `order.unpublishApporder()`, `order.unpublishDatasetorder()`, `order.unpublishWorkerpoolorder()`, `order.unpublishRequestorder()`
 
 ### Removed